[xla:python] Add support for stateful FFI calls registered via Python.
The FFI API supports stateful custom calls, but this typically requires that a custom type be registered with the FFI and its "type id" be specified. This change adds a `custom_type_ids` argument to `register_custom_call_target` function which adds support for this feature via Python.

The API that I settled on was to explicitly pass the required `XLA_FFI_TypeId*` pointers via PyCapsules, the same way we handle the function pointers:

```
xla_client.register_custom_call_target(
    ..., custom_type_ids={"custom_type": encapsulated_type_id})
```

where `encapsulated_type_id` is defined something like:

```
nb::capsule(reinterpret_cast<void*>(&MyCustomType::id))
```

An alternative approach would be to use the new metadata extension to extract the custom types when a handler is registered, and register any custom types then. This might be a more user friendly API, but we'd need to think a bit about how to handle duplicate registrations, and I wasn't sure I was up to the required template metaprogramming!

PiperOrigin-RevId: 670935323